+++ src/Engine/nd/ramio/RamioGuy.h
@@ -94,6 +94,6 @@ public:
 	void move(int direction, float floatPixels);
 
 
-	virtual void render(float alpha, BobTexture* texture, BobTexture* shadowTexture) override;
+	virtual void render(float alpha, std::shared_ptr<BobTexture> texture, std::shared_ptr<BobTexture> shadowTexture) override;
 };